Implement proper #include in preprocessor
This patch implements #include directives in the preprocessor.
The following changes are made:
- The command line interface is changed so include files are
  no longer provided explicitly. The '-I' option can be used
  to add to the include search path as with a regular C compiler.
- #include is supported (including nested #includes)
- Line numbers from preprocessed files are mapped to the
  appropriate include file if they came from an include file,
  and hence should print naturally.

To implement this reliably, reusing the parsed header results
had to be removed. This penalizes the current performance of
the first phase by about 2x, but in reality, on the usual test
set the first phase compilation time goes from 0.5s to 1s. Using
multithreading this goes back down to about 0.6s, which I think
is well worth it for the friendlier interface. It can also likely
be improved by some internal caching if needs be.