Reworked handling of lexical scopes (Namespace)
This patch reworks the implementation of lexical scope construction and
variable name lookup (see the comments for the class VScope).
Declarations are now collected separately on a pass before the AST is
built. This allows us to extract the logic and have a better separated
interface (hence the scope building is now in a separate file - VScope),
and in general scopes can be added easily at arbitrary nodes (more on
that in a later patch).

Statically allocated names with a multiplicity greater than 1 are now
renamed such that the emitted Verilog identifier will contain the line
number of the Alogic definition as a suffix, instead of appending 'Dup'.
This allows for easier debugging of waveforms from Verilog simulation.

Also factored out BaseVisitor from AstBuilder (required for VScope)

Performance impact is minimal